+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
     - Revised report text, added figure captions and table titles (#172).
     - Streamlined the Python code responsible for generating the report (#175, #177).
 - GRCh38 coordinates are now mandatory in marker definitions (#176).
+- Jijna2 sub-templates added to handle differences in the report for single ended versus paired end reads (#179). 
 
 ### Fixed
 - Bug with inconsistent sorting of read counts for interlocus balance (#159).